+++ main.rb
@@ -3,5 +3,4 @@ def main
   app = App.new
   app.run
 end
-
 main
